added sort on pickTeamActivity

+++ OlympicsApp/app/src/main/java/com/toneloc/olympicsapp/Country.java
@@ -26,7 +26,7 @@ public class Country {
         return mName;
     }
 
-    public float getmPrice() {
+    public int getmPrice() {
         return mPrice;
     }
